Add Debug, Move var Declarations Within For Loop
Moving declarations in for loop means that the vars are actually changing now...
Still needs some debugging as the console log is indicating draws when the game function indicates losses.

Need to add more console.table debugging for each loop in order to see cpuSelection, playerSelection, and result for each loop.